ai_consulting/requirements.txt
crewai==0.1.20
langchain==0.1.0
python-dotenv==1.0.0
openai==1.3.0
pypdf==3.17.1
